🎡 Minor refactoring
