+++ api/the_spymaster_api/client.py
@@ -28,66 +28,11 @@ from .structs.duet.responses import (
     DuetNextMoveResponse,
     DuetStartGameResponse,
 )
+from .client_classic import ClassicGameClient
+from .client_duet import DuetGameClient
+from .structs import SERVICE_ERRORS
 
 log = logging.getLogger(__name__)
-
-
-class ClassicClient(HTTPClient):
-    def __init__(self, game_api_url: str, retry_strategy: Retry | None = DEFAULT_RETRY_STRATEGY):
-        super().__init__(
-            base_url=f"{game_api_url}/classic",
-            retry_strategy=retry_strategy,
-            common_errors=SERVICE_ERRORS,
-        )
-
-    def start_game(self, request: ClassicStartGameRequest) -> ClassicStartGameResponse:
-        data = self.post(endpoint="start/", data=request.model_dump())
-        return ClassicStartGameResponse(**data)
-
-    def clue(self, request: ClueRequest) -> ClassicClueResponse:
-        data = self.post(endpoint="clue/", data=request.model_dump())
-        return ClassicClueResponse(**data)
-
-    def guess(self, request: GuessRequest) -> ClassicGuessResponse:
-        data = self.post(endpoint="guess/", data=request.model_dump())
-        return ClassicGuessResponse(**data)
-
-    def next_move(self, request: NextMoveRequest) -> ClassicNextMoveResponse:
-        data = self.post(endpoint="next-move/", data=request.model_dump())
-        return ClassicNextMoveResponse(**data)
-
-    def get_game_state(self, request: GetGameStateRequest) -> ClassicGetGameStateResponse:
-        data = self.get(endpoint="state/", data=request.model_dump())
-        return ClassicGetGameStateResponse(**data)
-
-
-class DuetClient(HTTPClient):
-    def __init__(self, game_api_url: str, retry_strategy: Retry | None = DEFAULT_RETRY_STRATEGY):
-        super().__init__(
-            base_url=f"{game_api_url}/duet",
-            retry_strategy=retry_strategy,
-            common_errors=SERVICE_ERRORS,
-        )
-
-    def start_game(self, request: DuetStartGameRequest) -> DuetStartGameResponse:
-        data = self.post(endpoint="start/", data=request.model_dump())
-        return DuetStartGameResponse(**data)
-
-    def clue(self, request: ClueRequest) -> DuetClueResponse:
-        data = self.post(endpoint="clue/", data=request.model_dump())
-        return DuetClueResponse(**data)
-
-    def guess(self, request: GuessRequest) -> DuetGuessResponse:
-        data = self.post(endpoint="guess/", data=request.model_dump())
-        return DuetGuessResponse(**data)
-
-    def next_move(self, request: NextMoveRequest) -> DuetNextMoveResponse:
-        data = self.post(endpoint="next-move/", data=request.model_dump())
-        return DuetNextMoveResponse(**data)
-
-    def get_game_state(self, request: GetGameStateRequest) -> DuetGetGameStateResponse:
-        data = self.get(endpoint="state/", data=request.model_dump())
-        return DuetGetGameStateResponse(**data)
 
 
 class TheSpymasterClient(HTTPClient):
@@ -100,6 +45,8 @@ class TheSpymasterClient(HTTPClient):
         )
         self.classic = ClassicClient(game_api_url=game_api_url, retry_strategy=retry_strategy)
         self.duet = DuetClient(game_api_url=game_api_url, retry_strategy=retry_strategy)
+        self.classic = ClassicGameClient(game_api_url=game_api_url, retry_strategy=retry_strategy)
+        self.duet = DuetGameClient(game_api_url=game_api_url, retry_strategy=retry_strategy)
 
     def load_models(self, request: LoadModelsRequest) -> LoadModelsResponse:
         data = self.post(endpoint="load-models/", data=request.model_dump())
